+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - enable updating a component's peerDependencies after import using the component's package.json
 - show npm-client's warnings when they are about missing peer-dependencies
 - support passing arguments/flags to the package-manager by specifying them after '--' (e.g. bit import -- --no-optional)
+- fix outdated to print correct version numbers
 
 ## [0.12.6-dev.5] - 2018-02-20